Fix import/export for Assignable model
