+++ test/integration/ggrc/converters/test_import_helpers.py
@@ -319,6 +319,9 @@ class TestGetObjectColumnDefinitions(TestCase):
         "Test Plan",
         "Audit",
         "Owner",
+        "Assessor",
+        "Verifier",
+        "Creator",
         "Primary Contact",
         "Secondary Contact",
         "Assessment URL",
